+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `create:worker` command
+- `show:worker` command
+- `list:workers` command
+- `delete:worker` command
+- `reboot:worker` command
 
 ## [0.1.4] - 2017-06-11
 ### Changed